AddCaseInput: restructure as dir + cleanup
